template/pyproject.toml
[tool.poetry]
name = "template"
version = "0.1.0"
description = ""
authors = ["Hattomo (TomohiroHattori) <45988993+Hattomo@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = { url = "https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp310-cp310-linux_x86_64.whl"}
hydra-core = "^1.3.2"
mlflow = "^2.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
yapf = "^0.40.1"
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
